+++ src/main/java/com/atlassian/db/replica/api/circuitbreaker/DualConnectionException.java
@@ -1,7 +1,0 @@
-package com.atlassian.db.replica.api.circuitbreaker;
-
-public class DualConnectionException extends RuntimeException {
-    public DualConnectionException(String message, Throwable cause) {
-        super(message, cause);
-    }
-}