+++ src/main/java/com/atlassian/db/replica/internal/circuitbreaker/BreakerHandler.java
@@ -2,7 +2,6 @@ package com.atlassian.db.replica.internal.circuitbreaker;
 
 import com.atlassian.db.replica.api.SqlCall;
 import com.atlassian.db.replica.spi.circuitbreaker.CircuitBreaker;
-import com.atlassian.db.replica.api.circuitbreaker.DualConnectionException;
 
 import java.sql.SQLException;
 
@@ -13,21 +12,21 @@ public class BreakerHandler {
         this.breaker = breaker;
     }
 
-    public <T> T handle(SqlCall<T> call) {
+    public <T> T handle(SqlCall<T> call) throws SQLException {
         try {
             return call.call();
         } catch (Throwable throwable) {
             breaker.handle(throwable);
-            throw new DualConnectionException("Db replica call failed. CircuitBreaker is " + breaker.getState(), throwable);
+            throw throwable;
         }
     }
 
-    public void handle(SqlRunnable call) {
+    public void handle(SqlRunnable call) throws SQLException {
         try {
             call.run();
         } catch (Throwable throwable) {
             breaker.handle(throwable);
-            throw new DualConnectionException("Db replica call failed. CircuitBreaker is " + breaker.getState(), throwable);
+            throw throwable;
         }
     }